Split the manifest layer by purity instead of by crate
Moving workspace and dependency discovery between crates never settled
anything: wherever it went, it dragged a filesystem question with it, and
`wado-manifest` ended up needing a two-tier wasm contract — works on
wasip1, merely compiles on wasm32-unknown-unknown — to describe a crate
the browser playground and the VS Code server both link.

The line that matters is purity, not crate boundaries. `wado-manifest`
says where a dependency *belongs*: coordinates, versions, cache paths,
member globs, all derivable from a manifest and an already-parsed lock.
Finding out what is actually on disk is the filesystem host's job, so it
moves into `wado-lsp/src/host/discovery.rs`, beside the
`FilesystemCompilerHost` whose `dependency_index` is its only reason to
exist. `wado-cli` keeps importing it from there — it is borrowing that
host's answer, and it already depends on the crate.

`wado-manifest` is pure again, so it works on every target and its
contract needs no qualification. The reads that cannot run in a browser
now sit inside a host a browser never instantiates, which is where the
old `DependencyProvider` seam already said this kind of thing goes.

This reverts the crate extraction and the wasip1 contract; the smell it
was aimed at was module organisation inside `wado-lsp`, and my original
report over-diagnosed it as a crate-placement problem.
