+++ mise.toml
@@ -52,8 +52,7 @@ set -xe -o pipefail
 cargo check
 cargo check -p wado-compiler --target wasm32-unknown-unknown
 cargo check -p wado-lsp --target wasm32-unknown-unknown
-cargo check -p wado-manifest --target wasm32-wasip1
-cargo check -p wado-manifest --target wasm32-wasip2
+cargo check -p wado-manifest --target wasm32-unknown-unknown
 """
 
 [tasks.check-grammar]
